backend/server/requirements.txt
Django==3.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
djongo==1.3.6
pymongo==4.6.0
dnspython==2.4.2
python-dotenv==1.0.0
sqlparse==0.2.4
